[update]deleteOneメソッドの改修
・articleDaoJdbcImplクラス
・articleDaoクラス
・articleServiceクラス
の引数をLong型へ変更

+++ src/main/java/com/example/demo/login/domain/repository/ArticleDao.java
@@ -18,5 +18,5 @@ public interface ArticleDao {
 
     public int updateOne(Article article) throws DataAccessException;
 
-    public int deleteOne(int id) throws DataAccessException;
+    public int deleteOne(Long id) throws DataAccessException;
 }
