+++ src/main/java/com/example/demo/login/domain/service/ArticleService.java
@@ -54,7 +54,7 @@ public class ArticleService {
         return result;
     }
 
-    public boolean delete(int id) {
+    public boolean delete(Long id) {
 
         int rowNumber = dao.deleteOne(id);
